feat: watch only components
